% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Odds ratio}
\usage{
odds_ratio(a_case_exp, b_ctrl_exp, c_case_uex, d_ctrl_uex)
}
\arguments{
\item{a_case_exp}{\verb{<num>} Number of cases exposed.}

\item{b_ctrl_exp}{\verb{<num>} Number of controls exposed.}

\item{c_case_uex}{\verb{<num>} Number of cases unexposed.}

\item{d_ctrl_uex}{\verb{<num>} Number of controls unexposed.}
}
\value{
Odds ratio.
}
\description{
Odds ratio (OR):
\itemize{
\item \verb{OR = (Odds in the exposed)/(Odds in the unexposed)}
}
}
\examples{
odds_ratio(240,160,160,240)
}
