% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{risk_ratio_attr}
\alias{risk_ratio_attr}
\title{Risk ratio (relative risk)}
\usage{
risk_ratio_attr(a_caex, b_ncex, c_caux, d_ncux)
}
\arguments{
\item{a_caex}{\verb{<num>} Number of cases exposed.}

\item{b_ncex}{\verb{<num>} Number of non-cases exposed.}

\item{c_caux}{\verb{<num>} Number of cases unexposed.}

\item{d_ncux}{\verb{<num>} Number of non-cases unexposed.}
}
\value{
Risk ratio (relative risk) and/or attributable risk.
}
\description{
Incidence exposed (Ie) or Risk exposed (Re)):
\itemize{
\item \code{Re = cases_exposed/(cases_exposed + cases_exposed)}
}

Incidence unexposed (Iu) or risk unexposed (Ru):
\itemize{
\item \code{Ru = cases_unexposed/(cases_unexposed + non-cases_unexposed)}
}

Risk ratio (RR) or relative risk:
\itemize{
\item \code{RR = Incidence_exposed / Incidence_unexposed}
}

Attributable risk (AR) or risk difference:
\itemize{
\item \code{AR = Incidence_exposed - Incidence_unexposed}
}
}
\examples{
risk_ratio_attr(60,140,20,180)
risk_ratio_attr(60,140,20,180)
}
