% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Odds ratio}
\usage{
odds_ratio(a_caex, b_coex, c_caux, d_coux)
}
\arguments{
\item{a_caex}{\verb{<num>} Number of cases exposed.}

\item{b_coex}{\verb{<num>} Number of controls exposed.}

\item{c_caux}{\verb{<num>} Number of cases unexposed.}

\item{d_coux}{\verb{<num>} Number of controls unexposed.}
}
\value{
Odds ratio.
}
\description{
Odds ratio (OR):
\itemize{
\item \verb{OR = (Odds in the exposed)/(Odds in the unexposed)}
}
}
\examples{
odds_ratio(240,160,160,240)
}
