% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{risk_ratio_attr}
\alias{risk_ratio_attr}
\title{Risk ratio (relative risk)}
\usage{
risk_ratio_attr(a_case_exp, b_ncas_exp, c_case_uex, d_ncas_uex)
}
\arguments{
\item{a_case_exp}{\verb{<num>} Number of cases exposed.}

\item{b_ncas_exp}{\verb{<num>} Number of non-cases exposed.}

\item{c_case_uex}{\verb{<num>} Number of cases unexposed.}

\item{d_ncas_uex}{\verb{<num>} Number of non-cases unexposed.}
}
\value{
A list containing risk exposed, risk unexposed,
risk ratio (relative risk) and attributable risk.
}
\description{
Incidence exposed (Ie) or Risk exposed (Re):
\itemize{
\item \code{Re = cases_exposed/(cases_exposed + cases_exposed)}
}

Incidence unexposed (Iu) or risk unexposed (Ru):
\itemize{
\item \code{Ru = cases_unexposed/(cases_unexposed + non-cases_unexposed)}
}

Risk ratio (RR) or relative risk:
\itemize{
\item \code{RR = Incidence_exposed / Incidence_unexposed}
}

Attributable risk (AR) or risk difference:
\itemize{
\item \code{AR = Incidence_exposed - Incidence_unexposed}
}
}
\examples{
risk_ratio_attr(60,140,20,180)
risk_ratio_attr(60,140,20,180)
}
